move `updateReadMe` to a separate file
